backend/pyproject.toml
[tool.poetry]
name = "auto-match-backend"
version = "0.1.0"
description = "AI-based Resume & Job Matching System Backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.3"
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sentence-transformers = "^2.2.2"
torch = "^2.1.0"
transformers = "^4.35.0"
numpy = "^1.24.3"
scikit-learn = "^1.3.2"
PyMuPDF = "^1.23.7"
python-docx = "^1.1.0"
PyPDF2 = "^3.0.1"
openai = "^1.3.5"
python-dotenv = "^1.0.0"
httpx = "^0.25.1"
aiofiles = "^23.2.1"
python-dateutil = "^2.8.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

